Revert "Добавлено: lua более не используется"
