--- YobaSolarSystem/CelestialObject.cpp
@@ -1,18 +1,0 @@
-//
-// Created by dimka on 6/10/2020.
-//
-
-#include "CelestialObject.h"
-//sf::Texture CelestialObject::tex1 = sf::Texture();
-//sf::Texture CelestialObject::tex2 = sf::Texture();
-//sf::Texture CelestialObject::tex3 = sf::Texture();
-
-CelestialObject::CelestialObject()
-{
-//    tex1.loadFromFile("yoba3.png");
-//    tex2.loadFromFile("yoba2.png");
-//    tex3.loadFromFile("yoba4.png");
-    acceleration = { 0, 0 };
-    r = 0;
-    m = 0;
-}
