--- YobaSolarSystem/World.cpp
@@ -1,5 +1,0 @@
-//
-// Created by dimka on 6/10/2020.
-//
-
-#include "World.h"
