+++ YobaSolarSystem/imgui.ini
@@ -1,0 +1,10 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][ObjEditor]
+Pos=1056,128
+Size=430,450
+Collapsed=0
+
