+++ YobaSolarSystem/init_lua_script.lua
@@ -1,0 +1,9 @@
+
+world:addObject(CObject.new(vec2.new(300, 300), vec2.new(0, 0), 25, 1200))
+world:addObject(CObject.new(vec2.new(300, 980), vec2.new(3.1, 0), 15, 735))
+world:addObject(CObject.new(vec2.new(300, 590), vec2.new(2.9, 0), 10, 600))
+world:addObject(CObject.new(vec2.new(300,900),vec2.new(2,0),10,705))
+--world:addObject(CObject.new(vec2.new(300,-1200),vec2.new(3,-0.2),25,620))
+--world:addObject(CObject.new(vec2.new(300,-1500),vec2.new(3,-0.2),25,611))
+--world:addObject(CObject.new(vec2.new(300,-2000),vec2.new(3,-0.2),25,610))
+collision = false
